refactor: route segment
- id -> recruitId
