+++ src/pages/recruits/[recruitId]/applicants/index.tsx
