+++ src/stories/pages/recruit/RecruitDetailPage.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import type { ComponentType } from 'react';
 
-import RecruitDetailPage from '~/pages/recruits/[id]';
+import RecruitDetailPage from '~/pages/recruits/[recruitId]';
 import { PageLayout } from '~/stories/Layout';
 
 const meta: Meta<typeof RecruitDetailPage> = {
